+++ src/main/java/com/umc/TheGoods/domain/payment/Payment.java
@@ -1,0 +1,57 @@
+package com.umc.TheGoods.domain.payment;
+
+import com.umc.TheGoods.domain.common.BaseDateTimeEntity;
+import com.umc.TheGoods.domain.member.Member;
+import com.umc.TheGoods.domain.order.OrderItem;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+import static javax.persistence.FetchType.LAZY;
+
+@Entity
+@Table(name = "payment")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Payment extends BaseDateTimeEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "payment_id")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
+    private Member member;
+
+    @OneToOne(mappedBy = "payment", fetch = LAZY)
+    private OrderItem orderItem;
+
+    private Long price;
+
+    private PaymentStatus status;
+
+    private String paymentUid;  // 결제 고유 번
+
+    @Builder
+    public Payment(Long price, PaymentStatus status) {
+        this.price = price;
+        this.status = status;
+    }
+
+    public void changePaymentBySuccess(PaymentStatus status, String paymentUid) {
+        this.status = status;
+        this.paymentUid = paymentUid;
+    }
+
+
+    //==조회 로직==//
+
+    /**
+     * 전체 주문 가격 조회
+     */
+
+}
