+++ src/main/java/com/umc/TheGoods/domain/payment/PaymentStatus.java
@@ -1,0 +1,7 @@
+package com.umc.TheGoods.domain.payment;
+
+public enum PaymentStatus {
+    OK,
+    READY,
+    CANCEL
+}
