+++ src/main/java/com/umc/TheGoods/repository/payment/PaymentRepository.java
@@ -1,6 +1,6 @@
 package com.umc.TheGoods.repository.payment;
 
-import com.umc.TheGoods.domain.Payment;
+import com.umc.TheGoods.domain.payment.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PaymentRepository extends JpaRepository<Payment, Long> {
